feat: 补齐目录接入与部署入口
