+++ glama.json
@@ -1,0 +1,6 @@
+{
+	"$schema": "https://glama.ai/mcp/schemas/server.json",
+	"maintainers": [
+		"can4hou6joeng4"
+	]
+}
